# assets/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Shopping Cart API","description":"Shopping Cart RESTful API"},"paths":{"/api/v{version}/Cart":{"post":{"tags":["Cart"],"operationId":"ApiV{versionCartPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/CreateCartRequest"}},{"name":"version","in":"path","required":true,"type":"string"}],"responses":{"201":{"description":"Success"},"400":{"description":"Bad Request"}}}},"/api/v{version}/Cart/{cartId}":{"get":{"tags":["Cart"],"operationId":"ApiV{versionCartByCartIdGet","consumes":[],"produces":["application/json"],"parameters":[{"name":"cartId","in":"path","required":true,"type":"string","format":"uuid"},{"name":"version","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/CartResponse"}},"404":{"description":"Not Found"}}},"put":{"tags":["Cart"],"operationId":"ApiV{versionCartByCartIdPut","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"cartId","in":"path","required":true,"type":"string","format":"uuid"},{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/AdjustQuantityRequest"}},{"name":"version","in":"path","required":true,"type":"string"}],"responses":{"204":{"description":"Success"},"400":{"description":"Bad Request"},"404":{"description":"Not Found"}}},"post":{"tags":["Cart"],"operationId":"ApiV{versionCartByCartIdPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["application/json"],"parameters":[{"name":"cartId","in":"path","required":true,"type":"string","format":"uuid"},{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/AddItemRequest"}},{"name":"version","in":"path","required":true,"type":"string"}],"responses":{"201":{"description":"Success","schema":{"$ref":"#/definitions/CartResponse"}},"400":{"description":"Bad Request"},"404":{"description":"Not Found"}}},"delete":{"tags":["Cart"],"operationId":"ApiV{versionCartByCartIdDelete","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"cartId","in":"path","required":true,"type":"string","format":"uuid"},{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/RemoveItemRequest"}},{"name":"version","in":"path","required":true,"type":"string"}],"responses":{"204":{"description":"Success"},"400":{"description":"Bad Request"},"404":{"description":"Not Found"}}}},"/api/v{version}/Cart/{cartId}/clear":{"delete":{"tags":["Cart"],"operationId":"ApiV{versionCartByCartIdClearDelete","consumes":[],"produces":[],"parameters":[{"name":"cartId","in":"path","required":true,"type":"string","format":"uuid"},{"name":"version","in":"path","required":true,"type":"string"}],"responses":{"204":{"description":"Success"},"404":{"description":"Not Found"}}}}},"definitions":{"CreateCartRequest":{"type":"object","properties":{"customerId":{"format":"uuid","type":"string"}}},"CartResponse":{"type":"object","properties":{"cartId":{"format":"uuid","type":"string"},"customerId":{"format":"uuid","type":"string"},"totalPrice":{"format":"double","type":"number"},"items":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/CartItemResponse"}}}},"CartItemResponse":{"type":"object","properties":{"productName":{"type":"string"},"price":{"format":"double","type":"number"},"quantity":{"format":"int32","type":"integer"}}},"AddItemRequest":{"type":"object","properties":{"productName":{"type":"string"},"price":{"format":"double","type":"number"},"quantity":{"format":"int32","type":"integer"}}},"AdjustQuantityRequest":{"type":"object","properties":{"productName":{"type":"string"},"quantity":{"format":"int32","type":"integer"}}},"RemoveItemRequest":{"type":"object","properties":{"productName":{"type":"string"}}}}}